[SofaKernel] REFACTOR remove the dependency between logging and Base.h
The problem: some of feature implemented in helper/logging rely on passing
a Base object. But the sofaHelper is not supposed to be linked with
sofaFramework.

To partially solve the problem, now only a forward definition of Base is used while
all the logging feature that suppose a knowledge of Base are now in sofa/framework/sofa/core/logging
instead of sofa/helper/logging

TODO: alternative designs exists but they often suppose to change the Base class to
(like super-classing the Base class to implement a "Loggable" interface). For the
moment I prefer to let untouch.